Health and Mana bars

+++ Assets/Scripts/Enemy/EnemyData.cs
@@ -11,18 +11,18 @@ public class EnemyData : MonoBehaviour
     [SerializeField]
     private float speed { get; set; }
     [SerializeField]
-    private int damage { get; set; }
+    private float damage { get; set; }
     [SerializeField]
     private float rangeAttack { get; set; }
 
     public float GetHealthPoint() { return healthPoint; }
     public float GetSpeed() { return speed; }
-    public int GetDamage() { return damage; }
+    public float GetDamage() { return damage; }
     public float GetRangeAttack() { return rangeAttack; }
 
     public void SetHealthPoint(float healthPoint) { this.healthPoint = healthPoint; }
-    public void SetDamage(int damage) {  this.damage = damage; }
-    public void SetSpeed(int speed) { this.speed = speed; }
+    public void SetDamage(float damage) {  this.damage = damage; }
+    public void SetSpeed(float speed) { this.speed = speed; }
 
 
     public void TakeDamage(int damage)
